requirements.txt
Django == 4.2.1
psycopg2-binary == 2.9.9
whitenoise == 6.6.0
plotly == 5.24.1
pandas == 2.2.3
django_plotly_dash == 2.4.5
channels-redis == 4.2.1
dpd-static-support == 0.0.5
django-bootstrap5 == 24.3
django-bootstrap-v5 == 1.0.11
dash-bootstrap-components == 1.7.1
dash-mantine-components == 0.15.3
uvicorn == 0.34.0
gunicorn